Put the conversion to Json String into ffi and keep QueryResults internally
